requirements.txt
streamlit
sentence-transformers
faiss-cpu
pandas
torch
requests
